requirements.txt
gradio
requests
pypdf
faiss-cpu
sentence-transformers
langchain-core
langchain-text-splitters
numpy
